requirements.txt
dash==2.9.3
pandas
plotly
seaborn
matplotlib.pyplot
dash_bootstrap_components